Prettify all files to match mainnet

+++ contracts/dnsregistrar/DNSClaimChecker.sol
@@ -7,7 +7,6 @@ import "../dnssec-oracle/RRUtils.sol";
 import "@ensdomains/buffer/contracts/Buffer.sol";
 
 library DNSClaimChecker {
-
     using BytesUtils for bytes;
     using RRUtils for *;
     using Buffer for Buffer.buffer;
@@ -26,7 +25,11 @@ library DNSClaimChecker {
         buf.append("\x04_ens");
         buf.append(name);
 
-        for (RRUtils.RRIterator memory iter = data.iterateRRs(0); !iter.done(); iter.next()) {
+        for (
+            RRUtils.RRIterator memory iter = data.iterateRRs(0);
+            !iter.done();
+            iter.next()
+        ) {
             bool found;
             address addr;
             (addr, found) = parseRR(data, iter.rdataOffset);
@@ -38,9 +41,14 @@ library DNSClaimChecker {
         return (address(0x0), false);
     }
 
-    function parseRR(bytes memory rdata, uint idx) internal pure returns (address, bool) {
+    function parseRR(bytes memory rdata, uint256 idx)
+        internal
+        pure
+        returns (address, bool)
+    {
         while (idx < rdata.length) {
-            uint len = rdata.readUint8(idx); idx += 1;
+            uint256 len = rdata.readUint8(idx);
+            idx += 1;
 
             bool found;
             address addr;
@@ -53,19 +61,27 @@ library DNSClaimChecker {
         return (address(0x0), false);
     }
 
-    function parseString(bytes memory str, uint idx, uint len) internal pure returns (address, bool) {
+    function parseString(
+        bytes memory str,
+        uint256 idx,
+        uint256 len
+    ) internal pure returns (address, bool) {
         // TODO: More robust parsing that handles whitespace and multiple key/value pairs
         if (str.readUint32(idx) != 0x613d3078) return (address(0x0), false); // 0x613d3078 == 'a=0x'
         if (len < 44) return (address(0x0), false);
         return hexToAddress(str, idx + 4);
     }
 
-    function hexToAddress(bytes memory str, uint idx) internal pure returns (address, bool) {
+    function hexToAddress(bytes memory str, uint256 idx)
+        internal
+        pure
+        returns (address, bool)
+    {
         if (str.length - idx < 40) return (address(0x0), false);
-        uint ret = 0;
-        for (uint i = idx; i < idx + 40; i++) {
+        uint256 ret = 0;
+        for (uint256 i = idx; i < idx + 40; i++) {
             ret <<= 4;
-            uint x = str.readUint8(i);
+            uint256 x = str.readUint8(i);
             if (x >= 48 && x < 58) {
                 ret |= x - 48;
             } else if (x >= 65 && x < 71) {
